Added most of the the room messages and fileinfo types.

+++ MatrixSDK/MatrixSDK/Structures/MatrixEvent.cs
@@ -23,7 +23,7 @@ namespace MatrixSDK.Structures
 
 	public class MatrixEventUnsigned{
 		public MatrixEventUnsigned prev_content;
-		public int age;
+		public Int64 age;
 		public string transaction_id;
 	}
 	/// <summary>
